feat: 이메일 인증 코드 확인 기능 개발, 여러번 인증 코드를 전송해도 최종 인증 코드만 사용 가능

+++ src/main/java/com/Sucat/global/infra/email/service/EmailSendService.java
@@ -1,21 +1,24 @@
 package com.Sucat.global.infra.email.service;
 
+import com.Sucat.global.infra.email.exception.EmailException;
+import com.Sucat.global.infra.email.model.VerificationCode;
 import com.Sucat.global.infra.email.repository.VerificationCodeRepository;
-import com.Sucat.global.infra.email.dto.VerificationCode;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
-import java.util.NoSuchElementException;
 import java.util.Random;
 
+import static com.Sucat.global.common.code.ErrorCode.*;
 import static com.Sucat.global.common.constant.EmailConstants.EXPIRATION_TIME_IN_MINUTES;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class EmailSendService {
 
     @Value("${spring.mail.host}")
@@ -29,7 +32,7 @@ public class EmailSendService {
         mailMessage.setTo(to);
         mailMessage.setSubject(String.format("Email Verification For %s", to));
 
-        VerificationCode verificationCode = generateVerificationCode(sentAt);
+        VerificationCode verificationCode = generateVerificationCode(sentAt, to);
         verificationCodeRepository.save(verificationCode);
 
         String text = verificationCode.generateCodeMessage();
@@ -37,17 +40,22 @@ public class EmailSendService {
         mailSender.send(mailMessage);
     }
 
-    public void verifyCode(String code, LocalDateTime verifiedAt) {
-        VerificationCode verificationCode = verificationCodeRepository.findByCode(code)
-                .orElseThrow(NoSuchElementException::new);
+    public void verifyCode(String email, String code,LocalDateTime verifiedAt) {
+        VerificationCode verificationCode = verificationCodeRepository.findByEmail(email)
+                .orElseThrow(() -> new EmailException(INVALID_VERIFICATION_EMAIL));
+
         if (verificationCode.isExpired(verifiedAt)) {
-            throw new NoSuchElementException();
+            throw new EmailException(VERIFICATION_CODE_EXPIRED);
+        }
+
+        if (!verificationCode.getCode().equals(code)) {
+            throw new EmailException(INVALID_VERIFICATION_CODE);
         }
 
         verificationCodeRepository.remove(verificationCode);
     }
 
-    private VerificationCode generateVerificationCode(LocalDateTime sentAt) {
+    private VerificationCode generateVerificationCode(LocalDateTime sentAt, String to) {
 
         Random r = new Random();
         String code = "";
@@ -55,8 +63,10 @@ public class EmailSendService {
             code += Integer.toString(r.nextInt(10));
         }
 
+        log.info("이메일 인증 코드 발급: {}", code);
         return VerificationCode.builder()
                 .code(code)
+                .email(to)
                 .createAt(sentAt)
                 .expirationTimeInMinutes(EXPIRATION_TIME_IN_MINUTES)
                 .build();
